requirements.txt
streamlit
pypdf
faiss-cpu
google-generativeai
numpy
langchain
langgraph
langchain-community
python-dotenv
